+++ hardhat-config/external.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  external: {
+    contracts: [
+      {
+        artifacts: ['node_modules/@layerzerolabs/lz-evm-sdk-v2', 'node_modules/@layerzerolabs/test-devtools-evm-hardhat'],
+      },
+    ],
+  },
+};