+++ hardhat-config/networks.config.js
@@ -1,4 +1,4 @@
-// import { EndpointId } from "@layerzerolabs/lz-definitions";
+const {EndpointId} = require('@layerzerolabs/lz-definitions');
 
 module.exports = {
   networks: {
@@ -40,8 +40,7 @@ module.exports = {
       url: 'http://mainnet',
       live: true,
       tags: ['production', 'ethereum'],
-      // eid: EndpointId.ETHEREUM_V2_MAINNET,
-      // eid: 30101, // LayerZero V2 Endpoint ID
+      eid: EndpointId.ETHEREUM_V2_MAINNET,
       companionNetworks: {
         base: 'base',
         polygon: 'matic',
@@ -57,8 +56,7 @@ module.exports = {
       url: 'http://sepolia',
       live: true,
       tags: ['staging', 'ethereum'],
-      // eid: EndpointId.SEPOLIA_V2_TESTNET,
-      // eid: 40161, // LayerZero V2 Endpoint ID
+      eid: EndpointId.SEPOLIA_V2_TESTNET,
       companionNetworks: {
         base: 'basesepolia',
         polygon: 'amoy',
@@ -79,8 +77,7 @@ module.exports = {
       url: 'http://matic',
       live: true,
       tags: ['production', 'polygon'],
-      // eid: EndpointId.POLYGON_V2_MAINNET,
-      // eid: 30109, // LayerZero V2 Endpoint ID
+      eid: EndpointId.POLYGON_V2_MAINNET,
       companionNetworks: {
         ethereum: 'mainnet',
       },
@@ -92,8 +89,7 @@ module.exports = {
       url: 'http://amoy',
       live: true,
       tags: ['staging', 'polygon'],
-      // eid: EndpointId.AMOY_V2_TESTNET,
-      // eid: 40267, // LayerZero V2 Endpoint ID
+      eid: EndpointId.AMOY_V2_TESTNET,
       companionNetworks: {
         ethereum: 'sepolia',
       },
@@ -110,8 +106,7 @@ module.exports = {
       url: 'http://bsc',
       live: true,
       tags: ['production', 'bsc'],
-      // eid: EndpointId.BSC_V2_MAINNET,
-      // eid: 30102, // LayerZero V2 Endpoint ID
+      eid: EndpointId.BSC_V2_MAINNET,
       companionNetworks: {
         ethereum: 'mainnet',
         arbitrum: 'arb1',
@@ -124,8 +119,7 @@ module.exports = {
       url: 'http://bsctest',
       live: true,
       tags: ['staging', 'bsc'],
-      // eid: EndpointId.BSC_V2_TESTNET,
-      // eid: 40102, // LayerZero V2 Endpoint ID
+      eid: EndpointId.BSC_V2_TESTNET,
       companionNetworks: {
         ethereum: 'sepolia',
         arbitrum: 'arbsepolia',
@@ -142,8 +136,7 @@ module.exports = {
       url: 'http://okex',
       live: true,
       tags: ['production', 'okex'],
-      // eid: EndpointId.OKX_V2_MAINNET,
-      // eid: 30155, // LayerZero V2 Endpoint ID
+      eid: EndpointId.OKX_V2_MAINNET,
       companionNetworks: {
         ethereum: 'mainnet',
       },
@@ -155,8 +148,7 @@ module.exports = {
       url: 'http://okextest',
       live: true,
       tags: ['staging', 'okex'],
-      // eid: EndpointId.OKX_V2_TESTNET,
-      // eid: 40155, // LayerZero V2 Endpoint ID
+      eid: EndpointId.OKX_V2_TESTNET,
       companionNetworks: {
         ethereum: 'sepolia',
       },
@@ -172,8 +164,7 @@ module.exports = {
       url: 'http://arb1',
       live: true,
       tags: ['production', 'arbitrum'],
-      // eid: EndpointId.ARBITRUM_V2_MAINNET,
-      // eid: 30110, // LayerZero V2 Endpoint ID
+      eid: EndpointId.ARBITRUM_V2_MAINNET,
       companionNetworks: {
         ethereum: 'mainnet',
         bsc: 'bsc',
@@ -186,8 +177,7 @@ module.exports = {
       url: 'http://arbsepolia',
       live: true,
       tags: ['staging', 'arbitrum'],
-      // eid: EndpointId.ARBSEP_V2_TESTNET,
-      // eid: 40231, // LayerZero V2 Endpoint ID
+      eid: EndpointId.ARBSEP_V2_TESTNET,
       companionNetworks: {
         ethereum: 'sepolia',
         bsc: 'bsctest',
@@ -203,8 +193,7 @@ module.exports = {
       url: 'http://base',
       live: true,
       tags: ['production', 'base'],
-      // eid: EndpointId.BASE_V2_MAINNET,
-      // eid: 30184, // LayerZero V2 Endpoint ID
+      eid: EndpointId.BASE_V2_MAINNET,
       companionNetworks: {
         ethereum: 'mainnet',
       },
@@ -215,8 +204,7 @@ module.exports = {
       url: 'http://basesepolia',
       live: true,
       tags: ['staging', 'base'],
-      // eid: EndpointId.BASESEP_V2_TESTNET,
-      // eid: 40245, // LayerZero V2 Endpoint ID
+      eid: EndpointId.BASESEP_V2_TESTNET,
       companionNetworks: {
         ethereum: 'sepolia',
       },