+++ src/templates/token/ERC20/layerzero_bridging/OFTAdapterMintBurn_deploy.js
@@ -1,0 +1,26 @@
+const Contract_deploy = require('../../../Contract/deploy');
+const {getContractAddress, getNamedAccount} = require('../../../../helpers/templates');
+
+module.exports = function (erc20DeploymentName, delegate, options = {}) {
+  const deploymentName = `OFTAdapterMintBurn@4.0_${erc20DeploymentName}`;
+  const migration = Contract_deploy(deploymentName, {
+    ...options,
+    contract: 'OFTAdapterMintBurn',
+    args: [
+      {name: 'token', value: getContractAddress(erc20DeploymentName)},
+      {name: 'lzEndpoint', value: getContractAddress('EndpointV2')},
+      {name: 'delegate', value: getNamedAccount(delegate)},
+    ],
+  });
+  migration.dependencies = [`${erc20DeploymentName}_deploy`];
+  migration.skip = async (hre) => {
+    const {log} = hre.deployments;
+    const {live, name} = hre.network;
+    if (!live) {
+      log(`${deploymentName}: non-live network ${name}, skipping...`);
+      return true;
+    }
+    return false;
+  };
+  return migration;
+};